refactor: move email service (and tests) into separate `mail` package
